debug:end moving is found correctly.

+++ nodes/ICP_Data_Processor.py
@@ -1,6 +1,7 @@
 import numpy as np
 import rospy
 from pyquaternion import Quaternion
+import os
 
 
 class ICPDataProcessor:
@@ -11,7 +12,6 @@ class ICPDataProcessor:
         self.first_rotation_matrix_icp = None
         self.first_transform_icp = None
         self.transformed_icp = None
-        self.speeds = None
         self.distances_icp = None
         self.start_of_moving = None
         self.end_of_moving = None
@@ -51,15 +51,6 @@ class ICPDataProcessor:
         coordinates = np.concatenate(icp, axis=1)
         self.transformed_icp = inv_matrix @ coordinates - np.expand_dims(inv_matrix @ coordinates[:, 0], axis=1)
 
-    def calculate_speeds_icp(self):
-        if self.transformed_icp is None:
-            return None
-        if self.speeds is None:
-            distances_one_period = np.abs(self.transformed_icp.T[1:] - self.transformed_icp.T[:-1])
-            times_one_period = self.times_icp[1:] - self.times_icp[:-1]
-            speeds_xyz = distances_one_period / times_one_period.reshape(-1, 1)
-            self.speeds = np.linalg.norm(speeds_xyz, axis=1)
-
     def get_icp_topic(self):
         for topic_name, topic_info in self.bag.get_type_and_topic_info()[1].items():
             if "/icp_odom" in topic_name:
@@ -77,19 +68,25 @@ class ICPDataProcessor:
         return self.distances_icp
 
     def get_average_speed(self):
-        return np.sum(self.speeds) / len(self.speeds) if self.speeds is not None else None
+        if self.transformed_icp is None:
+            return None
+        distances_one_period = np.abs(self.transformed_icp.T[1:] - self.transformed_icp.T[:-1])
+        times_one_period = self.times_icp[1:] - self.times_icp[:-1]
+        speeds_xyz = distances_one_period / times_one_period.reshape(-1, 1)
+        speeds = np.linalg.norm(speeds_xyz, axis=1)
+        return np.sum(speeds) / len(speeds) if speeds is not None else None
 
     def get_start_and_end_of_moving(self):
-        if self.start_of_moving is None and self.end_of_moving is None:
-            self.calculate_speeds_icp()
-            if self.speeds is None:
-                return None, None
-            moving = np.where(self.speeds > 0.2)[0]
-            if len(moving) == 0:
-                return None, None
-            saved_times = self.times_icp
-            self.start_of_moving = saved_times[moving[0]]
-            self.end_of_moving = saved_times[moving[-1] + 1] if len(moving) < len(saved_times) else saved_times[moving[-1]]
+        if self.transformed_icp is None:
+            return None
+        if self.start_of_moving is None:
+            distances_one_period_xyz = np.abs(self.transformed_icp.T[1:] - self.transformed_icp.T[:-1])
+            distances_one_period = np.linalg.norm(distances_one_period_xyz, axis=1)
+            for idx in range(len(distances_one_period)):
+                if distances_one_period[idx] > 0.002 and self.start_of_moving is None:
+                    self.start_of_moving = self.times_icp[idx]
+                elif distances_one_period[idx] > 0.002:
+                    self.end_of_moving = self.times_icp[idx]
         return self.start_of_moving, self.end_of_moving
 
     def get_times_icp(self):
@@ -116,17 +113,21 @@ class ICPDataProcessor:
         self.matrices_icp = object_["matrices"]
 
     def save_class_object(self, output_folder):
-        with open(f"{output_folder}/.data_availability.txt", 'r', encoding="utf-8") as file:
-            lines = file.readlines()
         state_icp = "False"
         if self.transformed_icp is not None:
             state_icp = "True"
             np.savez(f"{output_folder}/.icp.npz", coordinates=self.transformed_icp, saved_times=self.times_icp,
                      first_matrix=self.first_rotation_matrix_icp, first_transform=self.first_transform_icp,
                      matrices=self.matrices_icp)
-        with open(f"{output_folder}/.data_availability.txt", 'w', encoding="utf-8") as file:
-            for line in lines:
-                if line.startswith('icp'):
-                    file.write(f"icp {state_icp}\n")
-                else:
-                    file.write(line)
+        if os.path.exists(f"{output_folder}/.data_availability.txt"):
+            with open(f"{output_folder}/.data_availability.txt", 'r', encoding="utf-8") as file:
+                lines = file.readlines()
+            with open(f"{output_folder}/.data_availability.txt", 'w', encoding="utf-8") as file:
+                for line in lines:
+                    if line.startswith('icp'):
+                        file.write(f"icp {state_icp}\n")
+                    else:
+                        file.write(line)
+        else:
+            with open(f"{output_folder}/.data_availability.txt", 'w', encoding="utf-8") as file:
+                file.write(f"icp {state_icp}\n")
